Bug 1154470 - refactor artifact handling into its own model
