+++ tests/log_parser/test_tasks.py
@@ -13,6 +13,8 @@ from django.utils.six import BytesIO
 
 from ..sampledata import SampleData
 
+from treeherder.model.derived import ArtifactsModel
+
 
 @pytest.fixture
 def jobs_with_local_log(initial_data):
@@ -156,8 +158,8 @@ def test_parse_mozlog_log(jm, initial_data, jobs_with_local_mozlog_log,
     assert len(fails) == 3
 
 
-def test_parse_talos_log(jm, initial_data, jobs_with_local_talos_log, sample_resultset,
-                         mock_send_request, mock_get_remote_content):
+def test_parse_talos_log(jm, test_project, initial_data, jobs_with_local_talos_log,
+                         sample_resultset, mock_send_request, mock_get_remote_content):
     """
     check that performance job_artifacts get inserted when running
     a parse_log task for a talos job
@@ -169,7 +171,8 @@ def test_parse_talos_log(jm, initial_data, jobs_with_local_talos_log, sample_res
     jm.store_job_data(jobs)
     jm.process_objects(1, raise_errors=True)
 
-    artifact_list = jm.get_performance_artifact_list(0, 10)
+    with ArtifactsModel(test_project) as artifacts_model:
+        artifact_list = artifacts_model.get_performance_artifact_list(0, 10)
     assert len(artifact_list) >= 1  # should parse out at least one perf artifact
