+++ tests/webapp/api/test_artifact_api.py
@@ -4,19 +4,21 @@
 
 import pytest
 from django.core.urlresolvers import reverse
+from treeherder.model.derived import ArtifactsModel
 
 xfail = pytest.mark.xfail
 
 
 # we don't have/need an artifact list endpoint.
 
-def test_artifact_detail(webapp, eleven_jobs_processed, sample_artifacts, jm):
+def test_artifact_detail(webapp, test_project, eleven_jobs_processed, sample_artifacts, jm):
     """
     test retrieving a single job from the jobs-detail
     endpoint.
     """
     job = jm.get_job_list(0, 1)[0]
-    artifact = jm.get_job_artifact_references(job["id"])[0]
+    with ArtifactsModel(test_project) as artifacts_model:
+        artifact = artifacts_model.get_job_artifact_references(job["id"])[0]
 
     resp = webapp.get(
         reverse("artifact-detail",
