+++ treeherder/webapp/api/artifact.py
@@ -4,8 +4,8 @@
 
 from rest_framework import viewsets
 from rest_framework.response import Response
-from treeherder.webapp.api.utils import (UrlQueryFilter, with_jobs,
-                                         oauth_required)
+from treeherder.webapp.api.utils import UrlQueryFilter, oauth_required
+from treeherder.model.derived import JobsModel, ArtifactsModel
 
 
 class ArtifactViewSet(viewsets.ViewSet):
@@ -13,21 +13,20 @@ class ArtifactViewSet(viewsets.ViewSet):
     """
     This viewset is responsible for the artifact endpoint.
     """
-    @with_jobs
-    def retrieve(self, request, project, jm, pk=None):
+    def retrieve(self, request, project, pk=None):
         """
         retrieve a single instance of job_artifact
         """
         filter = UrlQueryFilter({"id": pk})
 
-        objs = jm.get_job_artifact_list(0, 1, filter.conditions)
-        if objs:
-            return Response(objs[0])
-        else:
-            return Response("job_artifact {0} not found".format(pk), 404)
+        with ArtifactsModel(project) as artifactModel:
+            objs = artifactModel.get_job_artifact_list(0, 1, filter.conditions)
+            if objs:
+                return Response(objs[0])
+            else:
+                return Response("job_artifact {0} not found".format(pk), 404)
 
-    @with_jobs
-    def list(self, request, project, jm):
+    def list(self, request, project):
         """
         return a list of job artifacts
         """
@@ -45,16 +44,21 @@ class ArtifactViewSet(viewsets.ViewSet):
         offset = filter.pop("offset", 0)
         count = min(int(filter.pop("count", 10)), 1000)
 
-        objs = jm.get_job_artifact_list(offset, count, filter.conditions)
-        return Response(objs)
+        with ArtifactsModel(project) as artifacts_model:
+            objs = artifacts_model.get_job_artifact_list(
+                offset,
+                count,
+                filter.conditions
+            )
+            return Response(objs)
 
-    @with_jobs
     @oauth_required
-    def create(self, request, project, jm):
+    def create(self, request, project):
 
         job_guids = [x['job_guid'] for x in request.DATA]
-        job_id_lookup = jm.get_job_ids_by_guid(job_guids)
+        with JobsModel(project) as jobsModel, ArtifactsModel(project) as artifacts_model:
 
-        jm.load_job_artifacts(request.DATA, job_id_lookup)
+            job_id_lookup = jobsModel.get_job_ids_by_guid(job_guids)
+            artifacts_model.load_job_artifacts(request.DATA, job_id_lookup)
 
-        return Response({'message': 'Artifacts stored successfully'})
+            return Response({'message': 'Artifacts stored successfully'})
